use a hash for docker::run env param
Hashes are the proper way of handing env vars as they accurately represent what an environment is. An environment is a list of names and values associated with each name. A name cannot be duplicated.
Using an array allows duplication. You can have ['foo=bar','foo=baz'] which is a perfectly valid array, but not valid for environment variables.
Using a hash also allows one to merge 2 hashes together and get deterministic result on collisions.

This change doesn't remove support for arrays, they still work fine.

This is also prepwork for a commit which will be following.