Fix stopping containers with docker::run upstart
The previous upstart method did not stop containers when the upstart service was stopped. This is a result of https://github.com/dotcloud/docker/pull/3240

As a workaround for the issue, on start we write the container ID to a file, and then use a post-stop script to kill that container.