[feature] added script to check for new images
the script connects to docker registry getting the current id of the
supplied image and tag. it compares this id with the id of the currently
installed image + tag. the puppet module has been altered so that the
docker pull command only runs if a new image is available. the pull
command also runs when there is no image with the given tag available in
the registry so that the puppet run has a chance to fail.