[fix] containers are not removed on start
docker containers should be ephemeral and due to the 1:1 relation
between the container and its init script (or job, unit, etc.) one
should remove the container first, before starting it again. this is
especially true for named containers, as container names are unique.

this fix changes the scripts for SysVinit and Upstart so that they
remove the old container on service startup. containers are removed on
start, not on stop, so that operators have a chance to inspect the old
conainer (filesystem, logs, etc.) before starting a new one.