Fix package name default for Debian 7
Prior to this commit, using this module on Debian 7 "Wheezy" with all
of the defaults required manually passing in the package name
(lxc-docker) because the case was not handled in params.pp. The upstream
repo is enabled by default, so I'd suspect that the least surprise to
users would be that the default package name was the same as it is in
the upstream repo: lxc-docker.

This commit modifies params.pp to include better defaults for Debian
Wheezy and modifies the test case accordingly.