Improve docker::run service configuration
Added two new parameters:

The `service_prefix` parameter allows overriding the prefix that is
applied to the service resource and the init script.  This makes it
easier to integrate with existing modules that have specific service
names in place.

The second parameter 'manage_service' allows turning off the creation of
the service resource.  Again, this makes it easier to integate with
existing modules that already manage the resource.  The restart_service
parameter will continue to notify the service if it is found under a
matching name.