Fix dependency cycle with docker::run
The docker::run command when using systemd notifies the exec for
reload, but then the service requires this exec. If you have multiple
runs, this creates a dependency cycle. I added a new exec based on the
sanitised title of the run so sure it wastes cycles reloading multiple
times, but it doesn't fail.