Removing deprecated properties.
The `key_source` and `required_packages` properties are deprecated and
`key` should now be a hash instead of a string. I'm not sure the best
way to handle the required packages so I've moved them to an
`ensure_packages` call for now and made that a requirement of the apt
resource.