Add manager flag for docker-storage-setup
Fixes a race condition with docker-storage-setup sysconfig template, the
package and its service. Docker-storage-setup depends on sysconfig/docker-storage
and sysconfig/docker-storage-setup files to be in place before it runs. By default
it runs as a service before the docker service starts. Currently there is an
ordering issue because the docker-storage-setup service is not being managed by
the docker::service class but the sysconfig file is, this causes the
docker-storage-service to not startup.