Docker::Run must respect $::docker::service_name
The system unit file generated by docker::run adds a
hard coded dependency on 'docker.service'.
If the docker service name is different - for example when
using the redhat extra packages - the name is 'docker-latest'.

This changes the template to use $::docker::service_name
when setting a dependency in the unit file.