Allow overriding service_provider on docker::run
Its possible to override the `service_provider` when managing the docker
daemon in case the defaults in params.pp aren't correct, but not when
defining a container using `docker::run`. This change adds a new
`service_provider` parameter to the run defined type, so that containers
can also be managed in these situations.

Fixes garethr/garethr-docker#755